[refactor] 강의실 삭제 로직 보완

+++ resource-server/src/main/java/com/inhabas/api/domain/lecture/usecase/LectureServiceImpl.java
@@ -48,6 +48,9 @@ public class LectureServiceImpl implements LectureService {
         if (lecture.notModifiableBy(memberId))
             throw new LectureCannotModifiableException();
 
+        if (!lecture.canBeDeleted())
+            throw new LectureCannotModifiableException("이미 진행된 강의는 삭제할 수 없습니다.");
+
         repository.delete(lecture);
     }
